Add /V1.0/InvoicePDF to Swagger and Get-AutotaskAPIResource to download invoice PDFs
Edits to allow request of Invoice PDF file format from Autotask, which used to be documented in the Autotask REST API docs and seems to have been removed, but still works. Was not listed in Swagger file, and required some changes to Get-AutotaskAPIResource to add functionality. Bumped version to 1.2.3 to compile and run locally where this has been used for over a year successfully. PSGetModuleInfo.xml file also exists locally with new version number but has been added to .gitignore to avoid pushing with commit.